plugin-ext: late textmate grammar activation
There is a race condition when the layout is restored: if an editor was
opened, it will be restored and wait for some language activation event.
The issue with this was first seen with the builtin cpp extension: two
cpp grammars are contributed, but the first one is not the right one to
use on a cpp file. When an editor opened, the first bogus grammar is
loaded and activated, triggering bogus coloration.

This commit ensures that all grammars are more correctly contributed
before activating anything in the next event-loop tick.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>